+++ lib/bitcoin_cleaner/executes.rb
@@ -1,8 +1,7 @@
-module BitcoinTestnet
+module BitcoinCleaner
   class Executes
 
     def self.command(c)
-      Dir.chdir BitcoinTestnet.dir
       prepared_command = PreparesCommand.for_silence(c)
       system prepared_command
     end