+++ spec/bitcoin_testnet/vcr_integrator_spec.rb
@@ -1,13 +1,13 @@
 require "spec_helper"
 
-module BitcoinTestnet
+module BitcoinCleaner
   describe VcrIntegrator do
 
     describe ".start" do
-      it "sets the BitcoinTestnet.current_test to be the given test" do
+      it "sets the BitcoinCleaner.current_test to be the given test" do
         test = double(metadata: {})
         described_class.start(test)
-        expect(BitcoinTestnet.current_test).to eq(test)
+        expect(BitcoinCleaner.current_test).to eq(test)
       end
     end