+++ spec/spec_helper.rb
@@ -1,2 +1,6 @@
 require "bundler/setup"
-Bundler.require(:default)
+Bundler.require(:default, :development)
+
+SPEC_DIR = File.dirname(__FILE__)
+TMP_DIR = File.join(SPEC_DIR, "..", "tmp")
+FileUtils.mkdir_p(TMP_DIR)